pyproject.toml
[project]
name = "misterknew"
version = "0.1.0"
description = ""
authors = [
    {name = "bratxemss",email = "bratxems@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0.0"
dependencies = [
    "tenacity (>=9.1.2,<10.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain-core (>=0.3.66,<0.4.0)",
    "langchain-deepseek (>=0.1.3,<0.2.0)",
    "langgraph (>=0.5.0,<0.6.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "playwright (>=1.53.0,<2.0.0)",
    "pycryptodomex (>=3.23.0,<4.0.0)",
    "pywin32 (>=310,<311)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "pypiwin32 (>=223,<224)",
    "pypsexec (>=0.3.0,<0.4.0)",
    "transformers (>=4.53.1,<5.0.0)",
    "accelerate (>=1.8.1,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "torch (>=2.7.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
